Use setInterval instead of process.nextTick
Using process.nextTick as a kind of timer on every single tick pegs
the CPU at 100% when a stream is paused. This commit substitutes
calls to process.nextTick with a call to setInterval which is less
efficient but does not peg the CPU nearly as badly.

In my tests only one paused stream using process.nextTick was enough
to peg the CPU. However, using setInterval I could create up to
10,000 paused streams and use ~95% of the CPU (measured using top).
Of course, this is a worst case scenario.

Fixes #3